+++ src/dyscord/client/discord_client.py
@@ -246,8 +246,9 @@ class DiscordClient:
                 #     # Heartbeat
                 #     pass
 
-                # elif opcode == 7:
-                #     # Reconnect
+                elif opcode == 7:
+                    self._log.debug('OPCODE: RECONNECT')
+                    await self._handle_op_7(data)
                 #     pass
 
                 # elif opcode == 9:
@@ -287,6 +288,14 @@ class DiscordClient:
 
             await self._gateway_ws.send(json.dumps(data))
 
+    async def _handle_op_7(self, data):
+
+        self._log.critical(f'Try to connect to {gateway_uri}')
+
+        await self._reconnect()
+
+        self._log.critical('Opcode 7 handled.')
+
     async def _handle_op_10(self, data):
 
         if self._heartbeat_task is not None:
@@ -295,6 +304,37 @@ class DiscordClient:
         self._heartbeat_task = asyncio.create_task(self._heartbeat(data['d']['heartbeat_interval']))
 
         self._log.debug('Opcode 10 handled.')
+
+    async def _reconnect(self):
+        '''Send a reconnect message.'''
+
+        gateway_uri = (await API.get_gateway_bot(self.token))['url']
+
+        if self._listener_task is not None:
+            self._listener_task.cancel()
+            self._listener_task = None
+
+        if self._heartbeat_task is not None:
+            self._heartbeat_task.cancel()
+            self._heartbeat_task = None
+            self._last_heartbeat_ack = None
+            self._gateway_ws = None
+
+        self._listener_task = asyncio.create_task(self._web_socket_listener(gateway_uri))
+
+        while self._gateway_ws is None:
+            await asyncio.sleep(1)
+
+        data = {
+            'op': 6,
+            'd': {
+                'token': self.token,
+                'session_id': self.session_id,
+                'seq': self._sequence_number,
+            }
+        }
+        self._log.info('Sending reconnect.')
+        await self._gateway_ws.send(json.dumps(data))
 
     async def _identify(self):
         data = {